:recycle: Merge cms headers
